# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triangle Curling Spares API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/availability/league": {
      "post": {
        "tags": [
          "availability"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "leagueId": {
                    "type": "number"
                  },
                  "available": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "leagueId",
                  "available"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/availability/can-skip": {
      "post": {
        "tags": [
          "availability"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "canSkip": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "canSkip"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/members/{memberId}/availability": {
      "get": {
        "tags": [
          "availability"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "memberId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "canSkip": {
                      "type": "boolean"
                    },
                    "availableLeagues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "leagueId": {
                            "type": "number"
                          },
                          "leagueName": {
                            "type": "string"
                          },
                          "dayOfWeek": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "leagueId",
                          "leagueName",
                          "dayOfWeek"
                        ]
                      }
                    }
                  },
                  "required": [
                    "canSkip",
                    "availableLeagues"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/availability/league/{leagueId}/members": {
      "get": {
        "tags": [
          "availability"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "position",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "leagueId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": [
                          "null",
                          "string"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "email"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}